add application service, update domain model

+++ src/Template.Domain/OrderPayed.cs
@@ -4,8 +4,8 @@ namespace Template.Domain
 {
     public class OrderPayed : DomainEvent
     {
-        public int OrderId { get; }
+        public long OrderId { get; }
 
-        public OrderPayed(int orderId) => OrderId = orderId;
+        public OrderPayed(long orderId) => OrderId = orderId;
     }
 }
